Moved constants file to resolve import error
