+++ tests/test_graphics/test_3d_map_wave_front.py
@@ -11,7 +11,7 @@ else:
 
 
 def graphics_test() -> None:
-    from constants import DATA_PATH, TEST_DATA_PATH
+    from utility.constants import DATA_PATH, TEST_DATA_PATH
     import pyautogui
 
     # Pick Uniform Random Fill with Wave-Front
